+++ .idea/gradle.xml
@@ -12,6 +12,8 @@
             <option value="$PROJECT_DIR$/app" />
             <option value="$PROJECT_DIR$/design" />
             <option value="$PROJECT_DIR$/design/resources" />
+            <option value="$PROJECT_DIR$/shared" />
+            <option value="$PROJECT_DIR$/shared/pizza" />
             <option value="$PROJECT_DIR$/util" />
           </set>
         </option>
