+++ app/build.gradle.kts
@@ -54,6 +54,7 @@ android {
 dependencies {
     implementation(project(":util"))
     implementation(project(":design:resources"))
+    implementation(project(":shared:pizza"))
 
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.lifecycle.runtime.ktx)
