--- app/src/main/java/com/example/pizzashift2025/pizza_main/data/Converter.kt
@@ -1,70 +1,0 @@
-package com.example.pizzashift2025.pizza_main.data
-
-import com.example.pizzashift2025.pizza_main.data.network.model.CatalogItemDto
-import com.example.pizzashift2025.pizza_main.data.network.model.DoughDto
-import com.example.pizzashift2025.pizza_main.data.network.model.IngredientDto
-import com.example.pizzashift2025.pizza_main.data.network.model.SizeDto
-import com.example.pizzashift2025.pizza_main.data.network.model.ToppingDto
-import com.example.pizzashift2025.pizza_main.domain.model.CatalogItem
-import com.example.pizzashift2025.pizza_main.domain.model.Dough
-import com.example.pizzashift2025.pizza_main.domain.model.Ingredient
-import com.example.pizzashift2025.pizza_main.domain.model.Size
-import com.example.pizzashift2025.pizza_main.domain.model.Topping
-import com.example.pizzashift2025.util.Constants.BASE_URL
-import javax.inject.Inject
-
-class Converter @Inject constructor() {
-
-    fun catalogItemDtoToDomain(model: CatalogItemDto): CatalogItem {
-        return CatalogItem(
-            id = model.id,
-            name = model.name,
-            ingredients = model.ingredients.map { ingredientDtoToDomain(it) },
-            toppings = model.toppings.map { toppingsDtoToDomain(it) },
-            description = model.description,
-            sizes = model.sizes.map { sizesDtoToDomain(it) },
-            doughs = model.doughs.map { doughDtoToDomain(it) },
-            calories = model.calories,
-            protein = model.protein,
-            totalFat = model.totalFat,
-            carbohydrates = model.carbohydrates,
-            sodium = model.sodium,
-            allergens = model.allergens,
-            isVegetarian = model.isVegetarian,
-            isGlutenFree = model.isGlutenFree,
-            isNew = model.isNew,
-            isHit = model.isHit,
-            img = BASE_URL + model.img
-        )
-    }
-
-    private fun ingredientDtoToDomain(model: IngredientDto): Ingredient {
-        return Ingredient(
-            name = model.name,
-            cost = model.cost,
-            img = model.img
-        )
-    }
-
-    private fun toppingsDtoToDomain(model: ToppingDto): Topping {
-        return Topping(
-            name = model.name,
-            cost = model.cost,
-            img = model.img
-        )
-    }
-
-    private fun sizesDtoToDomain(model: SizeDto): Size {
-        return Size(
-            name = model.name,
-            price = model.price
-        )
-    }
-
-    private fun doughDtoToDomain(model: DoughDto): Dough {
-        return Dough(
-            name = model.name,
-            price = model.price
-        )
-    }
-}
