--- app/src/main/java/com/example/pizzashift2025/pizza_main/data/network/api/PizzaApiService.kt
@@ -1,10 +1,0 @@
-package com.example.pizzashift2025.pizza_main.data.network.api
-
-import com.example.pizzashift2025.pizza_main.data.network.model.CatalogResponseDto
-import retrofit2.http.GET
-
-interface PizzaApiService {
-
-    @GET("pizza/catalog")
-    suspend fun getAllPizzaList(): CatalogResponseDto
-}
