--- app/src/main/java/com/example/pizzashift2025/pizza_main/data/network/api/Result.kt
@@ -1,6 +1,0 @@
-package com.example.pizzashift2025.pizza_main.data.network.api
-
-sealed class Result<out T> {
-    data class Success<out T>(val data: T): Result<T>()
-    data class Error(val reason: String): Result<Nothing>()
-}
