--- app/src/main/java/com/example/pizzashift2025/pizza_main/data/network/model/DoughDto.kt
@@ -1,6 +1,0 @@
-package com.example.pizzashift2025.pizza_main.data.network.model
-
-data class DoughDto(
-    val name: String,
-    val price: Int
-)
