--- app/src/main/java/com/example/pizzashift2025/pizza_main/data/network/model/IngredientDto.kt
@@ -1,7 +1,0 @@
-package com.example.pizzashift2025.pizza_main.data.network.model
-
-data class IngredientDto(
-    val name: String,
-    val cost: Int,
-    val img: String
-)
