--- app/src/main/java/com/example/pizzashift2025/pizza_main/data/network/model/SizeDto.kt
@@ -1,6 +1,0 @@
-package com.example.pizzashift2025.pizza_main.data.network.model
-
-data class SizeDto(
-    val name: String,
-    val price: Int
-)
