--- app/src/main/java/com/example/pizzashift2025/pizza_main/data/network/model/ToppingDto.kt
@@ -1,7 +1,0 @@
-package com.example.pizzashift2025.pizza_main.data.network.model
-
-data class ToppingDto(
-    val name: String,
-    val cost: Int,
-    val img: String
-)
