--- app/src/main/java/com/example/pizzashift2025/pizza_main/domain/model/Dough.kt
@@ -1,6 +1,0 @@
-package com.example.pizzashift2025.pizza_main.domain.model
-
-data class Dough(
-    val name: String,
-    val price: Int
-)
