--- app/src/main/java/com/example/pizzashift2025/pizza_main/domain/model/Topping.kt
@@ -1,7 +1,0 @@
-package com.example.pizzashift2025.pizza_main.domain.model
-
-data class Topping(
-    val name: String,
-    val cost: Int,
-    val img: String
-)
