+++ shared/pizza/src/main/kotlin/com/example/pizzashift2025/shared/pizza/data/network/model/CatalogResponseDto.kt
@@ -1,0 +1,7 @@
+package com.example.pizzashift2025.shared.pizza.data.network.model
+
+data class CatalogResponseDto(
+    val success: Boolean,
+    val reason: String?,
+    val catalog: List<CatalogItemDto>
+)
