+++ shared/pizza/src/main/kotlin/com/example/pizzashift2025/shared/pizza/data/network/model/SizeDto.kt
@@ -1,0 +1,6 @@
+package com.example.pizzashift2025.shared.pizza.data.network.model
+
+data class SizeDto(
+    val name: String,
+    val price: Int
+)
