+++ shared/pizza/src/main/kotlin/com/example/pizzashift2025/shared/pizza/data/network/model/ToppingDto.kt
@@ -1,0 +1,7 @@
+package com.example.pizzashift2025.shared.pizza.data.network.model
+
+data class ToppingDto(
+    val name: String,
+    val cost: Int,
+    val img: String
+)
