+++ shared/pizza/src/main/kotlin/com/example/pizzashift2025/shared/pizza/di/RepositoryPizzaMainModule.kt
@@ -1,0 +1,15 @@
+package com.example.pizzashift2025.shared.pizza.di
+
+import com.example.pizzashift2025.shared.pizza.data.repository.PizzaMainRepositoryImpl
+import com.example.pizzashift2025.shared.pizza.domain.repository.PizzaMainRepository
+import dagger.Module
+import dagger.Provides
+
+@Module
+object RepositoryPizzaMainModule {
+
+    @Provides
+    fun providePizzaMainRepository(repositoryImpl: PizzaMainRepositoryImpl): PizzaMainRepository {
+        return repositoryImpl
+    }
+}
