+++ shared/pizza/src/main/kotlin/com/example/pizzashift2025/shared/pizza/domain/model/Dough.kt
@@ -1,0 +1,6 @@
+package com.example.pizzashift2025.shared.pizza.domain.model
+
+data class Dough(
+    val name: String,
+    val price: Int
+)
