+++ shared/pizza/src/main/kotlin/com/example/pizzashift2025/shared/pizza/domain/model/Size.kt
@@ -1,0 +1,6 @@
+package com.example.pizzashift2025.shared.pizza.domain.model
+
+data class Size(
+    val name: String,
+    val price: Int
+)
