+++ shared/pizza/src/main/kotlin/com/example/pizzashift2025/shared/pizza/domain/model/Topping.kt
@@ -1,0 +1,7 @@
+package com.example.pizzashift2025.shared.pizza.domain.model
+
+data class Topping(
+    val name: String,
+    val cost: Int,
+    val img: String
+)
