+++ shared/pizza/src/main/kotlin/com/example/pizzashift2025/shared/pizza/domain/usecase/GetAllPizzaUseCase.kt
@@ -1,0 +1,14 @@
+package com.example.pizzashift2025.shared.pizza.domain.usecase
+
+import com.example.pizzashift2025.shared.pizza.data.network.api.Result
+import com.example.pizzashift2025.shared.pizza.domain.model.CatalogItem
+import com.example.pizzashift2025.shared.pizza.domain.repository.PizzaMainRepository
+import javax.inject.Inject
+
+class GetAllPizzaUseCase @Inject constructor(
+    private val pizzaRepository: PizzaMainRepository
+) {
+    suspend operator fun invoke(): Result<List<CatalogItem>> {
+        return pizzaRepository.getAllPizza()
+    }
+}
